Run tests for recipes in the run_list only.
The minitest-chef-handler gem does now detect and run tests if a path is not
specified. It would be nice to remove the code in this cookbook that builds
the path altogether.

Unfortunately the gem also runs all tests rather than just those related to
recipes in the run_list. It also doesn't cope with cookbook_path being an array.